requirements.txt
fastapi==0.95.0
sqlmodel==0.0.19
SQLAlchemy==2.0.14
pydantic==1.10.16
uvicorn==0.20.0
passlib==1.7.4  # ¡Sin [bcrypt]!
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.2.1
httpx==0.27.0
python-multipart
